Optimized MainViewModel.kt's home fragments' live data sources.

+++ app/src/main/java/com/ekosoftware/financialpreview/data/model/CurrencyConversion.kt
@@ -2,8 +2,6 @@ package com.ekosoftware.financialpreview.data.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
-import androidx.room.Index
-import androidx.room.PrimaryKey
 import java.util.*
 
 @Entity(tableName = "currencies", primaryKeys = ["currencyCodeId"])
@@ -16,7 +14,7 @@ data class CurrencyConversion(
     var name: String = "",
 
     @ColumnInfo(name = "currencyFactor")
-    var factor: Double = 0.0,
+    var factor: Long = 0,
 
     @ColumnInfo(name = "currencyLastUpdate")
     var lastUpdate: Date?
