+++ app/src/main/java/com/ekosoftware/financialpreview/data/model/movement/Projection.kt
@@ -1,5 +1,3 @@
 package com.ekosoftware.financialpreview.data.model.movement
 
-import java.util.*
-
-data class Projection(val currency: String, val currentMonthBalance: Double, val sixMonthsBalance: Double, val yearBalance: Double)
+data class Projection(val currency: String, val currentMonthBalance: Long, val sixMonthsBalance: Long, val yearBalance: Long)
