+++ app/src/main/java/com/ekosoftware/financialpreview/data/model/movement/QuickViewSummary.kt
@@ -3,8 +3,8 @@ package com.ekosoftware.financialpreview.data.model.movement
 data class QuickViewSummary(
     val currencyCode: String,
     val months: List<String>,
-    val incomes: List<Double>,
-    val expenses: List<Double>,
-    val balances: List<Double>,
-    val accumulatedBalances: List<Double>
+    val incomes: List<Long>,
+    val expenses: List<Long>,
+    val balances: List<Long>,
+    val accumulatedBalances: List<Long>
 )
